added ability to run tests on firefox splitted driver_util: driver, driver_options, driver factory, for easier maintenance added driver_types fixture: to run test on different browsers
