+++ pytest.ini
@@ -2,9 +2,9 @@
 #run tests in a row local
 #addopts = -rA -v --html=test_results.html --browser-type=local --env=stag --driver=mac
 #run tests parallel with xdist
-#addopts = -rA -v --html=test_results.html --browser-type=local --env=stag -n 2
+#addopts = -rA -v --html=test_results.html --browser-type=firefox --env=stag -n 2
 #run tests in CI
-addopts = -rA -v --env=dev --type=local
+addopts = -rA -v --env=dev --type=firefox
 markers =
     smoke: run smoke
     regression: run regression tests
@@ -21,3 +21,8 @@ filterwarnings =
     ignore::ResourceWarning
     ignore::FutureWarning
     ignore::PendingDeprecationWarning
+
+[pytest.config]
+# Define a custom command line option for driver types as a list
+type =
+    local,firefox, remote: Run tests with both local and Firefox drivers
